Large commit integrating changes made in MASC internal repo
This commit integrates a large number of changes from the
ESESC repo in the MASC lab that were made during the Spring '14
quarter.

Some of the major changes are:

- Added the boost library to misc instead of requiring it as a
dependency (most lines of code added).

- Updates to MOESI coherence to fix bugs for some specific cases

- Updates to TLB and added it to the default configuration

- Many updates to GPU modeling (still an alpha version)

- Initial infrastructure to support building ESESC as a dynamic library

- Updates that fix bug with memory controller.  It is not enabled by
  default

- Updates to prefetcher code to integrate with new memory interface and
  fix some bugs.

- Add some unit test infrastructure for testing memory hierarchy
  simulation.

- Added simu.conf file to model exynos processor used in Chromebook.

+++ simu/libnet/Message.cpp
@@ -31,11 +31,11 @@ void Message::forwardMsgAbs(Time_t time , Router *r) {
 void Message::dump()
 {
   MSG("MESSAGE #%d: src[%d:%d] dst[%d:%d] size[%d] latency[%lld]"
-      ,msgID
+      ,(int)msgID
       ,srcRouterID,srcPortID
       ,dstRouterID,dstPortID
-      ,nSize
-      ,globalClock - launchTime
+      ,(int)nSize
+      ,(long long)(globalClock - launchTime)
       );
 }
